edit document and divide to 2 files
