+++ solutions/hidden_element_solution.py
@@ -5,7 +5,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 
-from stale_element_lib import *
+from hidden_element_lib import *
 
 import time
